# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AITeachMe",
    "description": "本地优先的 AI 助教后端服务。",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/subjects/{subject}/chats/send": {
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "发送消息",
        "description": "保留原生 SSE 返回。",
        "operationId": "send_chat_api_v1_subjects__subject__chats_send_post",
        "parameters": [
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSendRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE 事件流。",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "请求参数不合法。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务内部异常。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "上游模型调用失败。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "依赖服务当前不可用。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subjects/{subject}/chats/list": {
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "聊天记录列表",
        "operationId": "list_chat_api_api_v1_subjects__subject__chats_list_post",
        "parameters": [
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatListRequest",
                "default": {
                  "page": 1,
                  "size": 20
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PaginatedData_ChatMessageItem__"
                }
              }
            }
          },
          "400": {
            "description": "请求参数不合法。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务内部异常。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subjects/{subject}/chats/clear": {
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "清空聊天记录",
        "operationId": "clear_chat_api_api_v1_subjects__subject__chats_clear_post",
        "parameters": [
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatClearRequest",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ChatClearData_"
                }
              }
            }
          },
          "400": {
            "description": "请求参数不合法。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务内部异常。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subjects/{subject}/chats/sessions/list": {
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "会话列表",
        "operationId": "list_chat_sessions_api_api_v1_subjects__subject__chats_sessions_list_post",
        "parameters": [
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSessionListRequest",
                "default": {
                  "page": 1,
                  "size": 20
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PaginatedData_ChatSessionItem__"
                }
              }
            }
          },
          "400": {
            "description": "请求参数不合法。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务内部异常。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subjects/{subject}/chats/threads/list": {
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "划词问答轮次列表",
        "operationId": "list_chat_threads_api_api_v1_subjects__subject__chats_threads_list_post",
        "parameters": [
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatThreadListRequest",
                "default": {
                  "page": 1,
                  "size": 20,
                  "source": "quick_chat"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PaginatedData_ChatThreadTurnItem__"
                }
              }
            }
          },
          "400": {
            "description": "请求参数不合法。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务内部异常。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subjects/{subject}/chats/sessions/create": {
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "创建会话",
        "operationId": "create_chat_session_api_api_v1_subjects__subject__chats_sessions_create_post",
        "parameters": [
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSessionCreateRequest",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ChatSessionCreateData_"
                }
              }
            }
          },
          "400": {
            "description": "请求参数不合法。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务内部异常。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subjects/{subject}/chats/sessions/delete": {
      "post": {
        "tags": [
          "chats"
        ],
        "summary": "删除会话",
        "operationId": "delete_chat_session_api_api_v1_subjects__subject__chats_sessions_delete_post",
        "parameters": [
          {
            "name": "subject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSessionDeleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ChatSessionDeleteData_"
                }
              }
            }
          },
          "400": {
            "description": "请求参数不合法。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务内部异常。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_ChatClearData_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务码，0 表示成功。",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息。",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatClearData"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据。"
          }
        },
        "type": "object",
        "title": "ApiResponse[ChatClearData]",
        "example": {
          "code": 0,
          "data": {},
          "message": "ok"
        }
      },
      "ApiResponse_ChatSessionCreateData_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务码，0 表示成功。",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息。",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatSessionCreateData"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据。"
          }
        },
        "type": "object",
        "title": "ApiResponse[ChatSessionCreateData]",
        "example": {
          "code": 0,
          "data": {},
          "message": "ok"
        }
      },
...
      "ApiResponse_PaginatedData_ChatMessageItem__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务码，0 表示成功。",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息。",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginatedData_ChatMessageItem_"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据。"
          }
        },
        "type": "object",
        "title": "ApiResponse[PaginatedData[ChatMessageItem]]",
        "example": {
          "code": 0,
          "data": {},
          "message": "ok"
        }
      },
      "ApiResponse_PaginatedData_ChatSessionItem__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务码，0 表示成功。",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息。",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginatedData_ChatSessionItem_"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据。"
          }
        },
        "type": "object",
        "title": "ApiResponse[PaginatedData[ChatSessionItem]]",
        "example": {
          "code": 0,
          "data": {},
          "message": "ok"
        }
      },
      "ApiResponse_PaginatedData_ChatThreadTurnItem__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务码，0 表示成功。",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息。",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginatedData_ChatThreadTurnItem_"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据。"
          }
        },
        "type": "object",
        "title": "ApiResponse[PaginatedData[ChatThreadTurnItem]]",
        "example": {
          "code": 0,
          "data": {},
          "message": "ok"
        }
      },
...
      "ChatClearRequest": {
        "properties": {
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Optional session ID to clear only one session."
          }
        },
        "type": "object",
        "title": "ChatClearRequest",
        "description": "Request body for clearing chat history.",
        "example": {
          "session_id": "optional-session-id"
        }
      },
...
      "ChatListRequest": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "description": "页码，从 1 开始。",
            "default": 1
          },
          "size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Size",
            "description": "每页数量。",
            "default": 20
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Optional session ID filter."
          }
        },
        "type": "object",
        "title": "ChatListRequest",
        "description": "Pagination request for chat history.",
        "example": {
          "page": 1,
          "size": 20
        }
      },
...
      "ChatSendRequest": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question",
            "description": "Current user question."
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Optional session ID. Auto-created when omitted."
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "description": "Optional source tag that enables direct chat mode."
          },
          "anchor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anchor Id",
            "description": "Optional doc heading anchor for highlighted QA."
          },
          "selected_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Context",
            "description": "Optional highlighted context."
          },
          "source_chunk_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Chunk Id",
            "description": "Optional source chunk ID for the highlighted context."
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "ChatSendRequest",
        "description": "Request body for sending one chat message.",
        "example": {
          "anchor_id": "chapter-1",
          "question": "什么是条件概率？",
          "selected_context": "条件概率表示在 B 已经发生的前提下 A 发生的概率。",
          "session_id": "dbe63613-08f6-4818-8317-cdf8d7a794a8",
          "source": "doc_selection",
          "source_chunk_id": 12
        }
      },
...
      "ChatSessionCreateRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Optional session title."
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "description": "Optional source tag."
          }
        },
        "type": "object",
        "title": "ChatSessionCreateRequest",
        "description": "Request body for creating a chat session."
      },
...
      "ChatSessionDeleteRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "Target session ID."
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "ChatSessionDeleteRequest",
        "description": "Request body for deleting one chat session."
      },
...
      "ChatSessionListRequest": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "description": "页码，从 1 开始。",
            "default": 1
          },
          "size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Size",
            "description": "每页数量。",
            "default": 20
          }
        },
        "type": "object",
        "title": "ChatSessionListRequest",
        "description": "Pagination request for chat sessions.",
        "example": {
          "page": 1,
          "size": 20
        }
      },
      "ChatThreadListRequest": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "description": "页码，从 1 开始。",
            "default": 1
          },
          "size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Size",
            "description": "每页数量。",
            "default": 20
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "description": "Optional source tag filter, defaults to doc quick chat.",
            "default": "quick_chat"
          }
        },
        "type": "object",
        "title": "ChatThreadListRequest",
        "description": "Pagination request for doc-selection chat turns.",
        "example": {
          "page": 1,
          "size": 20
        }
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务码，0 表示成功。",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息。",
            "default": "ok"
          },
          "data": {
            "type": "null",
            "title": "Data",
            "description": "业务数据。"
          }
        },
        "type": "object",
        "title": "ErrorResponse",
        "description": "统一错误响应结构。",
        "example": {
          "code": 404,
          "message": "资源不存在"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}